requirements.txt
Flask==1.1.4
flask-restplus==0.13.0
Flask-SQLAlchemy==2.1
requests==2.26.0
SQLAlchemy==1.4.23
urllib3==1.26.6
Werkzeug==0.16.1
Pillow==5.4.1
opencv-python==4.0.0.21